local-backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "denker-local-backend"
version = "0.1.0"
authors = [
  { name="Denker Team", email="dev@denker.ai" },
]
description = "Denker local backend server"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License", # Or your chosen license
    "Operating System :: OS Independent",
]
dependencies = [
    # Add your main dependencies from requirement-mcp.txt here
    # For example:
    "fastapi==0.115.12",
    "uvicorn==0.23.2",
    "pydantic==2.11.4",
    "mcp-agent==0.0.17",
    "fastmcp==2.2.6",
    "mcp==1.6.0",
    "mcp-server-fetch==2025.1.17",
    "mcp-server-qdrant==0.7.1",
    # ... and so on for all packages in requirement-mcp.txt
    # It's important for PyInstaller to know about these for script generation
    "aiohttp==3.11.18",
    "tenacity==9.1.2",
    "jinja2==3.1.6",
    "httpx==0.28.1",
    "markdownify==1.1.0",
    "protego==0.4.0",
    "readabilipy==0.3.0",
    "watchdog==6.0.0",
    "pypdf==4.3.1",
    "python-docx==1.1.2",
    "pandas==2.2.3",
    "pillow==11.2.1",
    "pytesseract==0.3.13",
    "beautifulsoup4==4.13.4",
    "markdown==3.8",
    "commonmark==0.9.1",
    "pyyaml==6.0.2",
    "openai==1.78.1",
    "anthropic==0.50.0",
    "langchain==0.3.25",
    "langchain-openai==0.3.16",
    "langchain-community==0.3.24",
    "langchain-core==0.3.59",
    "chromadb==0.6.3",
    "qdrant-client==1.14.2",
    "redis==6.1.0",
    "sentence-transformers==4.1.0",
    "sqlalchemy==2.0.40",
    "alembic==1.15.2",
    "psycopg2==2.9.10",
    "asyncpg==0.30.0",
    "unstructured[image,docx,xlsx,md,pdf]==0.17.2",
    "python-dotenv>=1.0.1",
    "pydantic-settings>=2.2.1"
]

[project.scripts]
denker-mcp-fetch = "mcp_server_fetch:main"
denker-mcp-qdrant = "mcp_server_qdrant.main:main"
denker-mcp-websearch = "mcp_local.servers.websearch.server:main"
denker-mcp-document-loader = "mcp_local.servers.document_loader.server:main"
denker-mcp-markdown-editor = "mcp_local.servers.markdown_editor.server:main"

[project.urls]
"Homepage" = "https://github.com/yourusername/denker-local-backend"
"Bug Tracker" = "https://github.com/yourusername/denker-local-backend/issues" 